Implement domain client

+++ Repositories/MicroORM/Memento/CodeCatalog.DDD.Memento/CodeCatalog.DDD.Data.Test.Integration/Helpers/OrderLineRow.cs
@@ -6,8 +6,8 @@ namespace CodeCatalog.DDD.Data.Test.Integration.Helpers
     {
         public long Id { get; set; }
         public ProductId ProductId { get; set; }
-        public double Discount { get; set; }
-        public double Price { get; set; }
+        public decimal Discount { get; set; }
+        public decimal Price { get; set; }
         public uint Quantity { get; set; }
     }
 }
